Create remote possibility.cpp
I wanted to build a RF and IR remote firmware that scans for the transmission from the remote instead of going through a predetermined list. This came about from a complaint from the Reddit user Count_vonDurban in the Cardputer forum about having troubles with a current program for this that got me thinking about this format instead. It allows you to save a remote file then store saved remote buttons in each one for different remotes. This is still untested, more to come.